Intel official release 2.3.15
The following changes were made in this release:
=====================================================
- add support for SLES16
- fix missing TX timestamps interrupts on E825 devices

+++ ice.spec
@@ -1,6 +1,6 @@
 Name: ice
 Summary: Intel(R) Ethernet Connection E800 Series Linux Driver
-Version: 2.3.14
+Version: 2.3.15
 Release: 1
 Source: %{name}-%{version}.tar.gz
 Vendor: Intel Corporation
@@ -64,27 +64,29 @@ make -C src
 %install
 make -C src INSTALL_MOD_PATH=%{buildroot} MANDIR=%{_mandir} modules_install_no_aux mandocs_install
 # Remove modules files that we do not want to include
-find %{buildroot}/lib/modules/ -name 'modules.*' -exec rm -f {} \;
+if [ -d %{buildroot}/lib/modules ]; then
+	find %{buildroot}/lib/modules/ -name 'modules.*' -exec rm -f {} \;
+fi
 cd %{buildroot}
-find lib -name "ice.ko" -printf "/%p\n" \
+find . -name "ice.ko" -printf "/%p\n" \
 	>%{_builddir}/%{name}-%{version}/file.list
-find lib -name "ice-vfio-pci.ko" -printf "/%p\n" \
+find . -name "ice-vfio-pci.ko" -printf "/%p\n" \
 	>>%{_builddir}/%{name}-%{version}/file.list
 %if (%need_aux_rpm == 2)
 make -C %{_builddir}/%{name}-%{version}/src INSTALL_MOD_PATH=%{buildroot} auxiliary_install
 
-find lib -path "*extern-symvers/intel_auxiliary.symvers" -printf "/%p\n" \
+find . -path "*extern-symvers/intel_auxiliary.symvers" -printf "/%p\n" \
 	>%{_builddir}/%{name}-%{version}/aux.list
-find * -name "auxiliary_bus.h" -printf "/%p\n" \
+find . -name "auxiliary_bus.h" -printf "/%p\n" \
 	>>%{_builddir}/%{name}-%{version}/aux.list
-find * -name "auxiliary_compat.h" -printf "/%p\n" \
+find . -name "auxiliary_compat.h" -printf "/%p\n" \
 	>>%{_builddir}/%{name}-%{version}/aux.list
-find * -name "kcompat_generated_defs.h" -printf "/%p\n" \
+find . -name "kcompat_generated_defs.h" -printf "/%p\n" \
 	>>%{_builddir}/%{name}-%{version}/aux.list
-find lib -name "intel_auxiliary.ko" -printf "/%p\n" \
+find . -name "intel_auxiliary.ko" -printf "/%p\n" \
 	>>%{_builddir}/%{name}-%{version}/aux.list
 %else
-find lib -name "intel_auxiliary.ko" -type f -delete
+find . -name "intel_auxiliary.ko" -type f -delete
 %endif
 
 export _ksrc=%{_usrsrc}/kernels/%{kernel_ver}
@@ -95,10 +97,9 @@ cd %{buildroot}
 %{!?privkey: %define privkey %{_sysconfdir}/pki/SECURE-BOOT-KEY.priv}
 %{!?pubkey: %define pubkey %{_sysconfdir}/pki/SECURE-BOOT-KEY.der}
 %{!?_signfile: %define _signfile ${_ksrc}/scripts/sign-file}
-for module in `find . -type f -name *.ko`;
-do
-strip --strip-debug ${module}
-$(KSRC=${_ksrc} %{_signfile} sha512 %{privkey} %{pubkey} ${module} > /dev/null 2>&1)
+for module in `find . -type f -name \*.ko`; do
+	strip --strip-debug ${module}
+	KSRC=${_ksrc} %{_signfile} sha512 %{privkey} %{pubkey} ${module}
 done
 %endif
